third_party/blink/web_tests/fast/events/wheel/wheelevent-in-scrolling-div.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../virtual/percent-based-scrolling/resources/percent-based-util.js"></script>
  <div id="overflow" style="border:2px solid black;overflow:auto;height:200px;width:200px;">
    <div style="background-color:red;height:200px;width:400px;"></div>
    <div style="background-color:green;height:200px;width:400px;"></div>
    <div style="background-color:red;height:200px;width:400px;"></div>
  </div>
  <div id="console"></div>
<script>
  const overflowElement = document.getElementById("overflow");
  overflowElement.addEventListener("mousewheel", mousewheelHandler, false);

  // scroll diagonally for 1 tick.
  const SCROLL_AMOUNT = pixelsPerTick();
  const PIXELS_TO_SCROLL = calculatePixelsToScroll(overflowElement, SCROLL_AMOUNT,
                                                   SCROLL_AMOUNT).y;
  const EXPECTED_WHEEL_DELTA = -PIXELS_TO_SCROLL / pixelsPerTick() *
                                LEGACY_MOUSE_WHEEL_TICK_MULTIPLIER *
                                window.devicePixelRatio;

  var eventHandlerResolve;
  eventHandlerPromise = new Promise(function(resolve) {
      eventHandlerResolve = resolve;
  });

  promise_test (async () => {
    var x = overflowElement.offsetLeft + 5;
    var y = overflowElement.offsetTop + 5
    await mouseMoveTo(x, y);
    await percentBasedSmoothScroll(SCROLL_AMOUNT, x, y, GestureSourceType.MOUSE_INPUT,
      "downright", SPEED_INSTANT, overflowElement);
    await waitForAnimationEndTimeBased(() => {
      return overflowElement.scrollTop;
    });
    await waitForAnimationEndTimeBased(() => {
      return overflowElement.scrollLeft;
    });

    assert_approx_equals(overflowElement.scrollTop, SCROLL_AMOUNT, 0.2);
    assert_approx_equals(overflowElement.scrollLeft, SCROLL_AMOUNT, 0.2);
    // The test will timeout if the event handler doesn't resolve this promise.
    await eventHandlerPromise;
  });

  function mousewheelHandler(e) {
    const approxeq = (a, b) => Math.abs(a-b) <= 0.2;
    // e.wheelDeltaX/Y is equal to number of ticks * 120. See kTickMultiplier
    // in src/third_party/blink/renderer/core/events/wheel_event.h
    if (approxeq(e.wheelDeltaY, EXPECTED_WHEEL_DELTA) &&
        approxeq(e.wheelDeltaX, EXPECTED_WHEEL_DELTA) &&
        approxeq(e.wheelDelta, EXPECTED_WHEEL_DELTA)) {
      eventHandlerResolve();
    }
  }
</script>
